-Modification fonction ajoutTrace -Ajout fonction affichage afficheTrace et afficheTrajet pour le debtggage -Ajout du fichier fonctionFront qui va contenir toutes les fonctions en commun entre la GUI front et le traitement Back

+++ Code/headers/structures.h
@@ -8,11 +8,12 @@ struct point
 };
 typedef struct point point;
 
+//Origine : coin haut-gauche
 struct rectangle
 {
      point     origine;
-     int       largeur;
-     int       hauteur;
+     double    largeur;
+     double    hauteur;
 };
 typedef struct rectangle rectangle;
